+++ src/main/java/blackjack/Application.java
@@ -1,0 +1,13 @@
+package blackjack;
+
+import blackjack.controller.BlackjackController;
+import blackjack.view.InputView;
+import blackjack.view.OutputView;
+
+public class Application {
+    public static void main(String[] args) {
+        BlackjackController blackjackController =
+                new BlackjackController(new InputView(), new OutputView());
+        blackjackController.run();
+    }
+}
