+++ src/main/java/blackjack/controller/GameCommand.java
@@ -1,0 +1,25 @@
+package blackjack.controller;
+
+import java.util.Arrays;
+
+public enum GameCommand {
+    PLAY("y"),
+    STOP("n");
+
+    private final String command;
+
+    GameCommand(String command) {
+        this.command = command;
+    }
+
+    public static GameCommand from(String input) {
+        return Arrays.stream(GameCommand.values())
+                     .filter(command -> command.command.equals(input))
+                     .findFirst()
+                     .orElseThrow(() -> new IllegalArgumentException("y 또는 n를 입력해주세요."));
+    }
+
+    public boolean isPlay() {
+        return this == PLAY;
+    }
+}
