+++ src/main/java/blackjack/domain/CardSuit.java
@@ -1,0 +1,18 @@
+package blackjack.domain;
+
+public enum CardSuit {
+    SPADE("스페이드"),
+    HEART("하트"),
+    DIAMOND("다이아몬드"),
+    CLUB("클로버");
+
+    private final String value;
+
+    CardSuit(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
