+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,0 +1,13 @@
+package blackjack.domain;
+
+public class Dealer extends Participant {
+
+    public Dealer() {
+        super(new Name("딜러"));
+    }
+
+    public boolean isHit() {
+        return this.getState()
+                   .isHit();
+    }
+}
