+++ src/main/java/blackjack/domain/Player.java
@@ -1,0 +1,12 @@
+package blackjack.domain;
+
+public class Player extends Participant {
+
+    public Player(String name) {
+        super(new Name(name));
+    }
+
+    public boolean isBust() {
+        return this.getState() == ScoreState.BUST;
+    }
+}
