--- src/main/java/domain/Dealer.java
@@ -1,13 +1,0 @@
-package domain;
-
-public class Dealer extends Participant {
-
-    public Dealer() {
-        super(new Name("딜러"));
-    }
-
-    public boolean isHit() {
-        return this.getState()
-                   .isHit();
-    }
-}
