--- src/main/java/domain/Player.java
@@ -1,12 +1,0 @@
-package domain;
-
-public class Player extends Participant {
-
-    public Player(String name) {
-        super(new Name(name));
-    }
-
-    public boolean isBust() {
-        return this.getState() == ScoreState.BUST;
-    }
-}
