+++ src/test/java/blackjack/domain/CardTest.java
@@ -1,0 +1,24 @@
+package blackjack.domain;
+
+import blackjack.domain.Card;
+import blackjack.domain.CardNumber;
+import blackjack.domain.CardSuit;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class CardTest {
+
+    @Test
+    @DisplayName("카드를 생성한다")
+    void createCardTest() {
+        Card card = new Card(CardSuit.HEART, CardNumber.ACE);
+
+        assertAll(
+                () -> assertThat(card.getSuit()).isEqualTo(CardSuit.HEART),
+                () -> assertThat(card.getNumber()).isEqualTo(CardNumber.ACE)
+        );
+    }
+}
